refactor: move some things around to prepare
