+++ src/bin/nix-index.rs
@@ -1,156 +1,32 @@
 //! Tool for generating a nix-index database.
-
-use error_chain::{error_chain, ChainedError};
+use error_chain::ChainedError;
 use separator::Separatable;
 use stderr::*;
 
 use clap::Parser;
-use futures::{future, FutureExt, Stream, StreamExt, TryFutureExt};
+use futures::{future, StreamExt};
 use std::ffi::OsStr;
 use std::fs::{self, File};
 use std::io::{self, Write};
 use std::iter;
 use std::sync::mpsc::channel;
 use std::thread;
-use std::iter::FromIterator;
 use std::path::PathBuf;
-use std::pin::Pin;
 use std::process;
 use std::str;
 
-use nix_index::database;
+use nix_index::database::Writer;
+use nix_index::errors::*;
 use nix_index::files::FileTree;
 use nix_index::hydra::Fetcher;
+use nix_index::listings::{fetch_file_listings, try_load_paths_cache};
 use nix_index::nixpkgs;
 use nix_index::package::StorePath;
-use nix_index::workset::{WorkSet, WorkSetHandle, WorkSetWatch};
 
 /// The URL of the binary cache that we use to fetch file listings and references.
 ///
 /// Hardcoded for now, but may be made a configurable option in the future.
 const CACHE_URL: &str = "http://cache.nixos.org";
-
-error_chain! {
-    errors {
-        QueryPackages {
-            description("query packages error")
-            display("querying available packages failed")
-        }
-        FetchFiles(path: StorePath) {
-            description("file listing fetch error")
-            display("fetching the file listing for store path '{}' failed", path.as_str())
-        }
-        FetchReferences(path: StorePath) {
-            description("references fetch error")
-            display("fetching the references of store path '{}' failed", path.as_str())
-        }
-        LoadPathsCache {
-            description("paths.cache load error")
-            display("loading the paths.cache file failed")
-        }
-        WritePathsCache {
-            description("paths.cache write error")
-            display("writing the paths.cache file failed")
-        }
-        CreateDatabase(path: PathBuf) {
-            description("crate database error")
-            display("creating the database at '{}' failed", path.to_string_lossy())
-        }
-        CreateDatabaseDir(path: PathBuf) {
-            description("crate database directory error")
-            display("creating the directory for the database at '{}' failed", path.to_string_lossy())
-        }
-        WriteDatabase(path: PathBuf) {
-            description("database write error")
-            display("writing to the database '{}' failed", path.to_string_lossy())
-        }
-        ParseProxy(err: nix_index::hydra::Error){
-            description("proxy parse error")
-            display("Can not parse proxy settings")
-        }
-    }
-}
-
-/// A stream of store paths (packages) with their associated file listings.
-///
-/// If a store path has no file listing (for example, because it is not built by hydra),
-/// the file listing will be `None` instead.
-type FileListingStream<'a> =
-    Pin<Box<dyn Stream<Item = Result<Option<(StorePath, FileTree)>>> + 'a>>;
-
-/// Fetches all the file listings for the full closure of the given starting set of path.
-///
-/// This function will fetch the file listings of each path in the starting set. Additionally, it
-/// will also determine the references of each path and recursively fetch the file listings for those
-/// paths.
-///
-/// The `jobs` argument is used to specify how many requests should be done in parallel. No more than
-/// `jobs` requests will be in-flight at any given time.
-fn fetch_file_listings(
-    fetcher: &Fetcher,
-    jobs: usize,
-    starting_set: Vec<StorePath>,
-) -> (FileListingStream, WorkSetWatch) {
-    // Create the queue that will hold all the paths that still need processing.
-    // Initially, only the starting set needs processing.
-    let workset = WorkSet::from_iter(starting_set.into_iter().map(|x| (x.hash().into_owned(), x)));
-
-    // Processes a single store path, fetching the file listing for it and
-    // adding its references to the queue
-    let process = move |mut handle: WorkSetHandle<_, _>, path: StorePath| {
-        fetcher
-            .fetch_references(path.clone())
-            .map_err(|e| Error::with_chain(e, ErrorKind::FetchReferences(path)))
-            .and_then(move |(path, references)| match references {
-                Some(references) => {
-                    for reference in references {
-                        let hash = reference.hash().into_owned();
-                        handle.add_work(hash, reference);
-                    }
-                    future::Either::Left(fetcher.fetch_files(&path).map(move |r| match r {
-                        Err(e) => Err(Error::with_chain(e, ErrorKind::FetchFiles(path))),
-                        Ok(Some(files)) => Ok(Some((path, files))),
-                        Ok(None) => Ok(None),
-                    }))
-                }
-                None => future::Either::Right(future::ok(None)),
-            })
-    };
-
-    // Process all paths in the queue, until the queue becomes empty.
-    let watch = workset.watch();
-    let stream = workset
-        .map(move |(handle, path)| process(handle, path))
-        .buffer_unordered(jobs);
-    (Box::pin(stream), watch)
-}
-
-/// Tries to load the file listings for all paths from a cache file named `paths.cache`.
-///
-/// This function is used to implement the `--path-cache` option.
-fn try_load_paths_cache() -> Result<Option<(FileListingStream<'static>, WorkSetWatch)>> {
-    let file = match File::open("paths.cache") {
-        Ok(file) => file,
-        Err(ref e) if e.kind() == io::ErrorKind::NotFound => return Ok(None),
-        Err(e) => return Err(e).chain_err(|| ErrorKind::LoadPathsCache)?,
-    };
-
-    let mut input = io::BufReader::new(file);
-    let fetched: Vec<(StorePath, FileTree)> =
-        bincode::deserialize_from(&mut input).chain_err(|| ErrorKind::LoadPathsCache)?;
-    let workset = WorkSet::from_iter(
-        fetched
-            .into_iter()
-            .map(|(path, tree)| (path.hash().to_string(), Some((path, tree)))),
-    );
-    let watch = workset.watch();
-    let stream = workset.map(|r| {
-        let (_handle, v) = r;
-        Ok(v)
-    });
-
-    Ok(Some((Box::pin(stream), watch)))
-}
 
 /// The main function of this module: creates a new nix-index database.
 async fn update_index(args: &Args) -> Result<()> {
@@ -249,7 +125,7 @@ async fn update_index(args: &Args) -> Result<()> {
     errst!("\r");
     fs::create_dir_all(&args.database)
         .chain_err(|| ErrorKind::CreateDatabaseDir(args.database.clone()))?;
-    let mut db = database::Writer::create(args.database.join("files"), args.compression_level)
+    let mut db = Writer::create(args.database.join("files"), args.compression_level)
         .chain_err(|| ErrorKind::CreateDatabase(args.database.clone()))?;
 
     let mut results: Vec<(StorePath, FileTree)> = Vec::new();
